updated shelf routes
