+++ views/shelves.pug
@@ -1,0 +1,18 @@
+extends layout.pug
+
+append head
+
+
+block content
+    h1= title
+    each shelf in shelves
+        .card.center
+            h2.card-title: a(href=`/shelves/${shelf.id}`)= shelf.name
+            each game in shelf.Games
+                p= game.title
+                img(src=game.imageURL, alt="")
+    form(action="/shelves", method="post")
+        input(type="hidden", name="_csrf", value=csrfToken)
+        label(for="name") Shelf Name
+        input(type="text", name="name", id="name")
+        button.btn.btn-primary(type="submit") Add Shelf
